requirements.txt
numpy==1.16.4
requests==2.18.4
matplotlib==3.1.0
Flask==1.0.3
redis==3.2.1
pandas==0.24.2
dask[dataframe]==0.19.1
scikit_learn==0.21.3
regex==2018.1.10
rq==1.1.0
gunicorn==19.9.0
google-cloud-storage==1.16.0
pyproj==1.9.6
